+++ scripts/build.sh
@@ -49,32 +49,18 @@ if [[ $_exit_code -ne 0 ]]; then
     exit 1
 fi
 
-echo "Generating style files"
-npm run compile-scss; _exit_code=$?
-if [[ $_exit_code -ne 0 ]]; then
-    echo "Failed to generate style files"
-    exit 1
-fi
-
 if [[ "$build_type" == "release" ]]; then # Compile everything for release mode
-    NODE_ENV=production npm run build-$build_target; _exit_code=$?
+    NODE_ENV=production npm run build-$build_target -- --env package=1; _exit_code=$?
     if [[ $_exit_code -ne 0 ]]; then
-        echo "Failed to build the $build_target applcation"
+        echo "Failed to build the $build_target application"
         exit 1
     fi
 elif [[ "$build_type" == "development" ]]; then
-    NODE_ENV=development npm run build-$build_target; _exit_code=$?
+    NODE_ENV=development npm run build-$build_target -- --env package=1; _exit_code=$?
     if [[ $_exit_code -ne 0 ]]; then
-        echo "Failed to build the $build_target applcation"
+        echo "Failed to build the $build_target application"
         exit 1
     fi
 fi
 
-echo "Generating environment"
-node file.js generate $build_target ${build_type}; _exit_code=$?
-if [[ $_exit_code -ne 0 ]]; then
-    echo "Failed to generate environment"
-    exit 1
-fi
-
-echo "$build_target builded successfully!"
+echo "$build_target build successfully!"