+++ scripts/install_dependencies.sh
@@ -34,52 +34,5 @@ install_node() {
     fi
 }
 
-install_rust() {
-    rustup_version=$(rustup --version 2>/dev/null)
-    if [[ $? -ne 0 ]]; then
-        echo "> Missing rustup, installing..."
-        curl https://build.travis-ci.org/files/rustup-init.sh -sSf | sh -s -- -y --default-toolchain nightly
-        # shellcheck disable=SC2181
-        [[ $? -ne 0 ]] && {
-            echo "> Failed to install rustup"
-            exit 1
-        }
-
-        PATH="$PATH:$HOME/.cargo/bin"
-        rustup_version=$(rustup --version 2>/dev/null)
-        echo "> Installed $rustup_version"
-    else
-        echo "> Found $rustup_version"
-    fi
-
-    echo "> Installing/updating the wasm32-unknown-unknown host"
-    rustup target add wasm32-unknown-unknown
-    if [[ $? -ne 0 ]]; then
-        echo "> Failed to install/updating the wasm target"
-        exit 1
-    fi
-}
-
-install_wasmpack() {
-    wasmpack_version=$(wasm-pack --version 2>/dev/null)
-    if [[ $? -ne 0 ]]; then
-        echo "> Missing wasm-pack, installing..."
-        curl https://rustwasm.github.io/wasm-pack/installer/init.sh -sSf | sh
-
-        # shellcheck disable=SC2181
-        [[ $? -ne 0 ]] && {
-            echo "> Failed to install wasm-pack"
-            exit 1
-        }
-
-        wasmpack_version=$(wasm-pack --version 2>/dev/null)
-        echo "> Installed $wasmpack_version"
-    else
-        echo "> Found $wasmpack_version"
-    fi
-}
-
 install_sys_deps
 install_node
-install_rust
-install_wasmpack