+++ shared/js/ui/frames/side/MusicPlaylistRenderer.tsx
@@ -13,6 +13,7 @@ import {preview_image} from "tc-shared/ui/frames/image_preview";
 import {joinClassList, useTr} from "tc-shared/ui/react-elements/Helper";
 import {spawnContextMenu} from "tc-shared/ui/ContextMenu";
 import {copyToClipboard} from "tc-shared/utils/helpers";
+import ImagePlaylistNoThumbnail from "./MusicPlaylistNoThumbnail.png";
 
 const cssStyle = require("./MusicPlaylistRenderer.scss");
 
@@ -58,7 +59,7 @@ export const DefaultThumbnail = (_props: { type: "loading" | "none-present" }) =
     return (
         <img
             draggable={false}
-            src="img/music/no-thumbnail.png"
+            src={ImagePlaylistNoThumbnail}
             alt={useTr("loading")}
         />
     );